# SwaggerInfo.yml
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation",
        "version": "1.0",
        "title": "Api Documentation",
        "termsOfService": "urn:tos",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "category-controller"
                ],
                "summary": "findAll",
                "operationId": "findAllUsingGET",
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Iterable«Category»"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "category-controller"
                ],
                "summary": "create",
                "operationId": "createUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "allErrors[0].arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "allErrors[0].code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "allErrors[0].codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "allErrors[0].defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "allErrors[0].objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "categoryData",
                        "description": "categoryData",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CategoryData"
                        }
                    },
                    {
                        "name": "errorCount",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "fieldError.arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldError.bindingFailure",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "fieldError.code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldError.defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.field",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.rejectedValue",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "fieldErrorCount",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "fieldErrors[0].arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldErrors[0].bindingFailure",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "fieldErrors[0].code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldErrors[0].defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].field",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].rejectedValue",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "globalError.arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalError.code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalError.codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalError.defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalError.objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalErrorCount",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "globalErrors[0].arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalErrors[0].code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalErrors[0].codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalErrors[0].defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalErrors[0].objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "nestedPath",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseData«Category»"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "category-controller"
                ],
                "summary": "update",
                "operationId": "updateUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "allErrors[0].arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "allErrors[0].code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "allErrors[0].codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "allErrors[0].defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "allErrors[0].objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "categoryData",
                        "description": "categoryData",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CategoryData"
                        }
                    },
                    {
                        "name": "errorCount",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "fieldError.arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldError.bindingFailure",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "fieldError.code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldError.defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.field",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldError.rejectedValue",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "fieldErrorCount",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "fieldErrors[0].arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldErrors[0].bindingFailure",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "fieldErrors[0].code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "fieldErrors[0].defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].field",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fieldErrors[0].rejectedValue",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "globalError.arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalError.code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalError.codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalError.defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalError.objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalErrorCount",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "globalErrors[0].arguments",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalErrors[0].code",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalErrors[0].codes",
                        "in": "query",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "globalErrors[0].defaultMessage",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "globalErrors[0].objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "nestedPath",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "objectName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseData«Category»"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "category-controller"
                ],
                "summary": "findOne",
                "operationId": "findOneUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                }
            },
            "title": "Category"
        },
        "CategoryData": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            },
            "title": "CategoryData"
        },
        "Iterable«Category»": {
            "type": "object",
            "title": "Iterable«Category»"
        },
...
        "ResponseData«Category»": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "payload": {
                    "$ref": "#/definitions/Category"
                },
                "status": {
                    "type": "boolean"
                }
            },
            "title": "ResponseData«Category»"
        },
...
    }
}